[로직 추가 개발, 리뷰 반영] 꿈나무 카드 관련 예외 처리 로직 추가 및 이전 리뷰 사항 반영

See merge request s11-fintech-finance-sub1/S11P21A601!146

+++ backend/card/src/main/java/kr/co/olivepay/card/service/Impl/CardTransactionServiceImpl.java
@@ -9,6 +9,7 @@ import kr.co.olivepay.card.repository.CardCompanyRepository;
 import kr.co.olivepay.card.repository.CardRepository;
 import kr.co.olivepay.card.service.CardTransactionService;
 import kr.co.olivepay.core.card.dto.req.CardSearchReq;
+import kr.co.olivepay.core.card.dto.res.enums.CardType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -29,8 +30,7 @@ public class CardTransactionServiceImpl implements CardTransactionService {
      * @param card
      * @return 등록된 카드 반환
      */
-    @Transactional
-    public Card registerCard(Card card) {
+    public Card registerCard(final Card card) {
         return cardRepository.save(card);
     }
 
@@ -40,8 +40,7 @@ public class CardTransactionServiceImpl implements CardTransactionService {
      * @param cardCompanyName
      * @return
      */
-    @Transactional(readOnly = true)
-    public CardCompany getCardCompany(String cardCompanyName) {
+    public CardCompany getCardCompany(final String cardCompanyName) {
         return cardCompanyRepository.findByName(cardCompanyName)
                                     .orElseThrow(() -> new AppException(ErrorCode.CARDCOMPANY_NOT_EXIST));
     }
@@ -52,8 +51,7 @@ public class CardTransactionServiceImpl implements CardTransactionService {
      * @param realCardNumber
      * @return
      */
-    @Transactional(readOnly = true)
-    public Optional<Card> getCard(String realCardNumber) {
+    public Optional<Card> getCard(final String realCardNumber) {
         Optional<Card> optionalCard = cardRepository.findByRealCardNumber(realCardNumber);
         return optionalCard;
     }
@@ -65,7 +63,7 @@ public class CardTransactionServiceImpl implements CardTransactionService {
      * @return
      */
     @Override
-    public Optional<Card> getDefaultCard(Long memberId) {
+    public Optional<Card> getDefaultCard(final Long memberId) {
         Optional<Card> optionalCard = cardRepository.findByMemberIdAndIsDefaultTrue(memberId);
         return optionalCard;
     }
@@ -78,10 +76,15 @@ public class CardTransactionServiceImpl implements CardTransactionService {
      * @param cardId
      */
     @Transactional
-    public void deleteCard(Long memberId, Long cardId) {
+    public void deleteCard(final Long memberId, final Long cardId) {
+        Card card = cardRepository.findById(cardId)
+                                  .orElseThrow(() -> new AppException(ErrorCode.CARD_NOT_EXIST));
+        if (card.getCardType() == CardType.DREAMTREE) {
+            throw new AppException(ErrorCode.DREAM_CARD_CAN_NOT_DELETE);
+        }
         int cnt = cardRepository.deleteByIdAndMemberId(cardId, memberId);
         if (cnt == 0) {
-            throw new AppException(ErrorCode.CARD_NOT_EXIST);
+            throw new AppException(ErrorCode.BAD_REQUEST);
         }
     }
 
@@ -92,7 +95,7 @@ public class CardTransactionServiceImpl implements CardTransactionService {
      * @return 해당 멤버의 모든 카드를 반환
      */
     @Transactional(readOnly = true)
-    public List<Card> getMyCardList(Long memberId) {
+    public List<Card> getMyCardList(final Long memberId) {
         return cardRepository.findByMemberIdOrderByIsDefaultDesc(memberId);
     }
 
@@ -105,13 +108,12 @@ public class CardTransactionServiceImpl implements CardTransactionService {
      * @return 꿈나무 카드, 차액 결제 카드, 공용 기부금 카드
      */
     @Transactional(readOnly = true)
-    public List<Card> getPaymentCardList(Long memberId, CardSearchReq cardSearchReq) {
+    public List<Card> getPaymentCardList(final Long memberId, CardSearchReq cardSearchReq) {
         return cardRepository.findByMemberIdAndCardSearchReq(memberId, cardSearchReq);
     }
 
     @Override
-    @Transactional(readOnly = true)
-    public Optional<Card> getCardWithAccountById(Long cardId) {
+    public Optional<Card> getCardWithAccountById(final Long cardId) {
         return cardRepository.findCardAndAccountById(cardId);
     }
